GenAI_HW1/requirements.txt
streamlit
openai
python-dotenv
PyMuPDF
python-docx